Fixed CodeGen issue with FP64 and function calls
This fixes issue #54.

The fix consists in removing the code that breaks f64 registers in 2x i32
ones, from get registers for calling convention and then break f64 values
only when needed, through RISCV lowering code.

The main problem with breaking f64 registers in
getRegisterTypeForCallingConv()/getNumRegistersForCallingConv() was that the
export registers would always be 2x i32 ones, one of them unallocated, and
it seemed that there was no reasonable way to change this without changing
the TargetLowering interface.
It could probably be easily solved if there was a way to tell CodeGen to
use different registers for exporting the returned value. But this seems
a very target specific corner case.